requirements.txt
python-chess==1.999
zstandard==0.23.0
